feat: add character slice
